requirements.txt
Flask~=3.1.1
Flask_SQLAlchemy

Werkzeug~=3.1.3
utils~=1.0.2
numpy~=2.3.1
pillow~=11.3.0